% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DendSer-package.R
\docType{package}
\name{DendSer-package}
\alias{DendSer-package}
\title{DendSer: Dendrogram Seriation: Ordering for Visualisation}
\description{
Re-arranges a dendrogram to optimize visualisation-based cost functions. The methods implemented here are desribed in Advances in Dendrogram Seriation for Application to Visualization, JCGS (2015) D. Earle and C.B. Hurley \doi{10.1080/10618600.2013.874295}.
}
\author{
\strong{Maintainer}: Catherine Hurley \email{catherine.hurley@mu.ie}

Authors:
\itemize{
  \item Catherine B. Hurley
  \item Denise Earle
}

}
\keyword{internal}
